Put everything in the one cpp file
Homework rating bot can't deal with a CMakeLists.txt file and a lib
directory here.
