--- homeworks/homework_3/guessing_game/lib/CMakeLists.txt
@@ -1,5 +1,0 @@
-add_library(myRandom STATIC myRandom.cpp)
-
-target_include_directories(myRandom
-    INTERFACE
-    $(CMAKE_CURRENT_SOURCE_DIR))
