--- homeworks/homework_3/guessing_game/lib/myRandom.h
@@ -1,3 +1,0 @@
-#pragma once
-
-int randomNumber(int lowerBound, int upperBound);
